Removed settings page, added 2FA to profile, removed chat bubble
