+++ front-vite/src/Layout/Main/index.tsx
@@ -3,7 +3,6 @@ import { Routes, Route } from 'react-router-dom';
 import Home from '../../Pages/Home/index';
 import Game from '../../Pages/Game/index';
 import ErrorPage from '../../Pages/Error/index';
-import UserSettings from '../../Pages/UserSettings/index';
 import ProfilePage from '../../Pages/Profile';
 import LoginPage from '../../Pages/Login/index';
 import ChannelsPage from '../../Pages/Channels/index';
@@ -23,7 +22,6 @@ export const Main: React.FC = () => {
       <Route path="/logout" element={<LogoutPage />} />
       <Route path="/game" element={<Game />} />
       <Route path="/channels" element={<ChannelsPage />} />
-      <Route path="/profile/settings" element={<UserSettings />} />
       <Route path="/profile/:username" element={<ProfilePage />} />
       <Route path="/404" element={<ErrorPage />} />
       <Route path="/viewusers" element={<AllUsersPage />} />
