+++ front-vite/src/Pages/Profile/index.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { Avatar, Box, Stack, Typography, useTheme, Divider, Grid, IconButton, Container } from '@mui/material';
+import { Avatar, Box, Stack, Typography, useTheme, Divider, Grid, IconButton, Container, Dialog, DialogTitle, DialogContent, TextField, DialogActions, Button, styled, FormControlLabel, Switch } from '@mui/material';
 import {Input} from '@mui/material'
 import { useMediaQuery, Tooltip } from '@mui/material';
 import { alpha } from '@mui/material/styles';
@@ -27,12 +27,14 @@ import { useUser,
 import { socket } from '../../Providers/NotificationContext/Notification';
 import { User, MatchData, MatchRatio } from '../../Types/User/Interfaces';
 import { UserStatus } from '../../Types/User/Enum';
+import axios from 'axios';
 
 const ProfilePage: React.FC = () => {
 	const theme = useTheme();
 	const navigate = useNavigate();
 	const isSmallScreen = useMediaQuery(theme.breakpoints.down('sm'));
-	const { user } = useUser();
+	const { user, setUser } = useUser();
+	const intraId = user.intraId;
 	const location = useLocation();
 	const pathSegments = location.pathname.split('/');
 	const lastSegment = pathSegments[pathSegments.length - 1]
@@ -51,7 +53,122 @@ const ProfilePage: React.FC = () => {
 	const [whichStatus, setWhichStatus] = useState<UserStatus>(UserStatus.Offline);
 	const [matchHistory, setMatchHistory] = useState<MatchData[]>([]);
 	const [ratioArr, setRatioArr] = useState<MatchRatio[]>([]);
+	// State
+    const [is2FAEnabled, setIs2FAEnabled] = useState(user?.twoFactorSecret || false);
+    const [qrCode, setQrCode] = useState<string | null>('');
+    const [secret, setSecret] = useState('');
+    // const [otp, setOtp] = useState('');
+    const [openQRDialog, setOpenQRDialog] = useState(false);
+    const [verificationCode, setVerificationCode] = useState('');
+    const [error, setError] = useState(false);
+
+	const check2FAStatus = async () => {
+		try {
+			const response = await axios.get(import.meta.env.URL_BACKEND_2FA_STATUS + `?intraId=${intraId}`);
+		} catch (error) {}
+	};
+
+	// Handle 2FA Toggle
+	const handle2FAToggle = async (event) => {
+		const enable2FA = event.target.checked;
+		
+		if (enable2FA) {
+			// User wants to enable 2FA → Generate QR Code
+			try {
+				const response = await axios.get(import.meta.env.URL_BACKEND_QR_GENERATE);
+				setQrCode(response.data.qrCode);
+				setSecret(response.data.secret)
+				setOpenQRDialog(true);
+			} catch (error) {
+				console.error('Error generating QR Code:', error);
+			}
+		} else {
+			// User wants to disable 2FA → Disable immediately
+			try {
+				const response = await axios.post(import.meta.env.URL_BACKEND_2FA_DELETE + `?intraId=${intraId}`);
+				setIs2FAEnabled(false);
+				check2FAStatus();
+			} catch (error) {}
+		}
+	};
 	
+	// Verify Code After Scanning QR
+	const handleVerifyQR = async () => {
+		try {
+			const response = await axios.post(import.meta.env.URL_BACKEND_QR_VERIFY, {
+				intraId,
+				secret,
+				token: verificationCode
+			});
+			check2FAStatus();
+			if (response.data.success) {
+				setIs2FAEnabled(true);
+				setUser({ ...user, twoFactorEnabled: true });
+				setOpenQRDialog(false);
+			} else {
+				setError(true);
+			}
+		} catch (error) {
+			console.error('Error verifying 2FA:', error);
+			setError(true);
+		}
+	}
+
+	const SettingsSection = styled(Box)(({ theme }) => ({
+		marginBottom: theme.spacing(4),
+	}));
+
+	const TwoFactorToggle = () => {
+		return (
+			<SettingsSection>
+				<br/>
+				<b>Two-Factor Authentication:</b>
+				<br/>
+				<FormControlLabel
+					control={<Switch checked={is2FAEnabled} onChange={handle2FAToggle} color='primary' />}
+				/>
+			</SettingsSection>
+		)
+	}
+	
+	const TwoFactorDialog = () => {
+		return (
+			<Dialog open={openQRDialog} onClose={() => setOpenQRDialog(false)}>
+				<DialogTitle>Scan QR Code</DialogTitle>
+				<DialogContent>
+					{qrCode ? <img src={qrCode} alt="2FA QR Code" style={{ width: '100%' }} /> : 'Loading QR Code...'}
+					<Typography variant="body2" style={{ marginTop: '1em' }}>
+						Enter the 6-digit code from your authentication app.
+					</Typography>
+					<TextField
+						fullWidth
+						error={error}
+						helperText={error ? 'Invalid verification code' : ''}
+						label="Verification Code"
+						variant="outlined"
+						margin="normal"
+						value={verificationCode}
+						onChange={(e) => {
+							setVerificationCode(e.target.value);
+							setError(false); // Reset error when new input comes in
+						}}
+						onKeyDown={(e) => {
+							if (e.key === 'Enter') {
+								handleVerifyQR();
+							}
+						}}
+						selectTextOnFocus // Highlights text when focused
+					/>
+	
+				</DialogContent>
+				<DialogActions>
+					<Button onClick={() => setOpenQRDialog(false)} color="secondary">Cancel</Button>
+					<Button onClick={handleVerifyQR} color="primary" variant="contained">Verify</Button>
+				</DialogActions>
+			</Dialog>
+		)
+	}
+
 	let RemoveFriend = (id:string) => {
 		var newlist: string[] = friendsList.filter((friend: String) => friend !== id);
 		setFriendsList(newlist);
@@ -774,6 +891,8 @@ const ProfilePage: React.FC = () => {
 					{userProfile.email}
 					<br />
 				</Typography>
+				{TwoFactorToggle()}
+				{TwoFactorDialog()}
 			</Stack>
 		);
 	}
@@ -783,8 +902,8 @@ const ProfilePage: React.FC = () => {
 		return (
 			<Container sx={{ padding: theme.spacing(3) }}>
 				<Stack
-						width={'100%'}
-						overflow={'hidden'}
+					width={'100%'}
+					overflow={'hidden'}
 				>
 					{userInfo()}
 					<Stack
@@ -804,6 +923,9 @@ const ProfilePage: React.FC = () => {
 	let getUserProfile = async () : Promise<void> =>
 	{
 		const tmp: User = await getUserFromDatabase(lastSegment, navigate);
+
+		if (!tmp)
+			return;
 
 		if (user.id == tmp.id)
 		{
@@ -835,7 +957,15 @@ const ProfilePage: React.FC = () => {
 			} 
 		});
 
-	}, [lastSegment]);
+		const fetch2FAStatus = async () => {
+			try {
+				const response = await axios.get(import.meta.env.URL_BACKEND_2FA_STATUS + `?intraId=${intraId}`);
+				setIs2FAEnabled(response.data.is2FAEnabled);
+			} catch (error) {}
+		};
+		fetch2FAStatus();
+
+	}, [lastSegment, intraId]);
 
 	let whichPage = () =>
 	{
